Move SmartSnapshotLeader outside PostgresConnectorTask to core
